feat: useNotification과 Notification 컴포넌트를 통해 뷰로직 분리
